refactor(analytics): split hedge nt ratio research helpers
